+++ ICD.Connect.Audio/Telemetry/VolumeTelemetryNames.cs
@@ -7,7 +7,11 @@ namespace ICD.Connect.Audio.Telemetry
 		public const string VOLUME_CONTROL_MUTE_COMMAND = "Volume Control Mute Command";
 
 		public const string VOLUME_CONTROL_PERCENT = "Volume Control Percent";
-		public const string VOLUME_CONTROL_VOLUME_CHANGED = "Volume Control Volume Changed";
+		public const string VOLUME_CONTROL_PERCENT_CHANGED = "Volume Control Volume Percent Changed";
 		public const string VOLUME_CONTROL_PERCENT_COMMAND = "Volume Control Percent Command";
+
+		public const string VOLUME_CONTROL_LEVEL = "Volume Control Level";
+		public const string VOLUME_CONTROL_LEVEL_CHANGED = "Volume Control Volume Level Changed";
+		public const string VOLUME_CONTROL_LEVEL_COMMAND = "Volume Control Level Command";
 	}
 }